+++ src/Exception/Business/BusinessException.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Anktx\Kafka\Client\Exception\Business;
-
-abstract class BusinessException extends \DomainException {}